+++ src/model/DealProductRequestBody.js
@@ -167,7 +167,7 @@ DealProductRequestBody.prototype['tax'] = 0;
 DealProductRequestBody.prototype['tax_method'] = undefined;
 
 /**
- * Whether the product is enabled for a deal or not. This makes it possible to add products to a deal with a specific price and discount criteria, but keep them disabled, which refrains them from being included in the deal value calculation. When omitted, the product will be marked as enabled by default
+ * Whether the product is enabled for a deal or not  This makes it possible to add products to a deal with a specific price and discount criteria, but keep them disabled, which refrains them from being included in the deal value calculation. When omitted, the product will be marked as enabled by default  Not possible to disable the product if the deal has installments associated and the product is the last one enabled  Not possible to enable the product if the deal has installments associated and the product is recurring 
  * @member {Boolean} enabled_flag
  * @default true
  */